алерт в универсальный алерт
